Implement new data component predicates in block predicates (#2732)
* Implement new data component predicates in block predicates

* Remove unused network type

* Replace `Bounds` with `Range`

* Uncomment and fix failing tests

* Encode the minecraft:lock data component as an item predicate

* Use an enum for component predicate types

* Fix ArmorTrim predicate returning false when materials match

* Allow tags in enchantment and potion data component predicates

The wiki says that tags should also be allowed in jukebox_playable and trim, but I don't think the game has any tags for those(?)

* Make CollectionPredicate accept a Collection instead of an Iterable

Not sure why I didn't do this initially

* Add nullability annotations

* Copy lists when necessary to make records truly immutable

* Ignore air when processing `container` predicate

* Remove duplication of codecs and `test` methods in Enchantments and StoredEnchantments predicates

* Use existing DataComponent.Holder inherited from ItemStack instead of constructing a new DataComponentMap

* Use NBTPredicate instead of requiring exact tag equality

Thanks GoldenStack for writing the NBT comparison method!

* Fix tests

* Remove unnecessary intermediate `Entry` record in AttributeModifiers predicate

* Use DataComponentMap's build-in serializer instead of custom one and fix BlockPredicate test with non-null but empty exact or partial matchers

* Fix serializing data component predicates as NBT

* Add more convenience constructors, create a CollectionPredicate Builder, and make some erroneously-private classes public

* Fix decoding data component predicates from NBT

* Fix testSingleBlockNbtInput

* Fix demo compile

* Add unit tests and Javadoc comments for DataComponentPredicate

* Rename componentPredicates -> components for consistency with other names

* Add more nullability annotations

* Convert ComponentPredicateType to a registry and simplify Codec and NetworkType implementations

* Remove static reference to registry in favor of MinecraftServer#componentPredicateTypes()

* Remove reference to deprecated TagStringIOExt

* Specify initial capacity when copying maps

* Make `min` and `max` in *Range classes optional to allow unbounded ends of ranges

* Add tests for NbtPredicate and fix one inconsistency with vanilla Minecraft's implementation

* Use Java Map type publicly, defensively copy the input map, and add method to remove item

* Remove some `@NotNull` annotations that I missed

* Implement various small suggestions from @kermandev

* Make `exact` and `predicates` non-null in DataComponentPredicates

* Add javadocs and Contract annotations to new Codec methods

* ComponentPredicateSet improvements:
- Convert ComponentPredicateSet to a regular (non-record) class to make private new no-copy constructor
- Add javadocs
- Remove unnecessary copy in add()
- Add new convenience constructor from a Collection<DataComponentPredicate>
- Make ComponentPredicateSet implement Predicate<DataComponent.Holder>

* fix: resolve compile and test errors after master merge

* refactor: define range codecs in Range classes and register villager variant predicate

* refactor: replace deprecated Check.notNull with Objects.requireNonNull and remove unused Check imports

* refactor: design ComponentPredicateSet as sealed interface with ComponentPredicateSetImpl record and revert NetworkBuffer mapValue changes

* fix: align component predicates with Minecraft 26.2

* refactor: Remove NBT_COMPOUND_COERCED

* fix: Bee

* feat: Proper tooltip

---------

+++ src/test/java/net/minestom/server/utils/entity/EntityFinderIntegrationTest.java
@@ -87,7 +87,7 @@ public class EntityFinderIntegrationTest {
         var far = spawnZombie(instance, 50);
 
         var finder = createFinder(EntityFinder.TargetSelector.ALL_ENTITIES)
-                .setDistance(new Range.Double(5, Float.MAX_VALUE));
+                .setDistance(new Range.Double(5.0, (double) Float.MAX_VALUE));
 
         assertEquals(List.of(far), finder.find(instance, null));
     }
@@ -111,7 +111,7 @@ public class EntityFinderIntegrationTest {
         var far = spawnZombie(instance, 50);
 
         var finder = createFinder(EntityFinder.TargetSelector.ALL_ENTITIES)
-                .setDistance(new Range.Double(-10, 5));
+                .setDistance(new Range.Double(-10.0, 5.0));
 
         assertEquals(List.of(near), finder.find(instance, null));
     }
